requirements.txt
mutagen==1.47.0
ruamel.yaml==0.18.10
termcolor==2.5.0
Wand==0.6.13
yt-dlp==2025.2.19
